added double prescision, and added sine osc tests

+++ include/IBDSP/common.hpp
@@ -2,5 +2,5 @@
 
 namespace IBDSP
 {
-    constexpr float TWO_PI = 6.283185307179586f;
+    constexpr double TWO_PI = 6.283185307179586f;
 }
