+++ include/IBDSP/oscillators/oscillator.hpp
@@ -11,7 +11,7 @@ namespace IBDSP::oscillators
         virtual void setSampleRate(float sampleRate) = 0;
         virtual void setAmplitude(float amp) = 0;
         virtual float process() = 0;
-        virtual void reset(float phase = 0.0f) = 0;
+        virtual void reset(double phase = 0.0f) = 0;
 
         // helper
         void processBlock(float* buffer, int numSamples)
